+++ src/ast/node.ts
@@ -1,9 +1,16 @@
-import { Operator } from "../lex/token"
+import { Operator, SoftKeyword } from "../lex/token"
+import { HzObjDecl, HzFuncDecl, HzVarDecl, HzNaryFuncDecl, getFuncSignature } from "../parse/declaration"
+import { HzFileDef } from "../parse/file"
+import { HzBreakStatmt, HzCodeBlock, HzContinueStatmt, HzExprStatmt, HzIfStatmt, HzInitStatmt, HzReturnStatmt, HzStatmt, HzWhileStatmt } from "../parse/statement"
+import { HzBinaryExpr, HzExpr, HzFuncCallExpr, HzLiteralExpr, HzNaryFuncCallExpr, HzNullaryFuncCallExpr, HzRefExpr } from "../parse/expr"
+import { LiteralType } from "../parse/literal"
 
 export type ASTNodeClass<T = unknown> = { new(...args: unknown[]): T }
 export abstract class ASTNode {
   parent?: ASTNode
   abstract get isLeaf(): boolean
+  *children(): Iterable<ASTNode> { }
+  build(): void { }
   findRef(name: string): RefNode | undefined {
     return this.parent?.findRef(name)
   }
@@ -18,6 +25,11 @@ export abstract class ASTNode {
 export class BlockNode extends ASTNode {
   locals: Map<string, RefNode> = new Map()
   statements: StatementNode[] = []
+  protected def: HzCodeBlock
+  constructor(def: HzCodeBlock) {
+    super()
+    this.def = def
+  }
   findRef(name: string): RefNode | undefined {
     return this.locals.get(name) ?? this.parent?.findRef(name)
   }
@@ -35,11 +47,30 @@ export class BlockNode extends ASTNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield* this.locals.values()
+    yield* this.statements.values()
+  }
+  build(): void {
+    for (const local of this.def.locals) {
+      this.defineLocal(new LocalVarNode(local.name))
+    }
+    for (const statmt of this.def.statements) {
+      const node = makeStatementNode(statmt)
+      this.addStatement(node)
+    }
+  }
 }
 
-export class FileNode extends BlockNode {
+export class FileNode extends ASTNode {
   declare parent: undefined
-  declare locals: Map<string, ObjNode | FuncNode | LocalVarNode>
+  locals: Map<string, ObjNode | FuncNode | LocalVarNode>
+  protected def: HzFileDef
+  statements: StatementNode[] = []
+  constructor(def: HzFileDef) {
+    super()
+    this.def = def
+  }
   findRef(name: string): ObjNode | FuncNode | LocalVarNode | undefined {
     return this.locals.get(name)
   }
@@ -49,21 +80,52 @@ export class FileNode extends BlockNode {
     symbol.parent = this
     this.locals.set(symbol.name, symbol)
   }
+  addStatement(statement: StatementNode): true {
+    this.statements.push(statement)
+    statement.parent = this
+    return true
+  }
   get isLeaf(): boolean {
     return false
+  }
+  *children(): Iterable<ASTNode> {
+    yield* this.locals.values()
+    yield* this.statements.values()
+  }
+  build(): void {
+    for (const topLevel of this.def.topLevels) {
+      if (topLevel instanceof HzObjDecl) {
+        this.defineLocal(new ObjNode(topLevel))
+      } else if (topLevel instanceof HzFuncDecl) {
+        this.defineLocal(new FuncNode(topLevel))
+      } else if (topLevel instanceof HzExprStatmt) {
+        this.addStatement(new ExprStatementNode(topLevel))
+      } else if (topLevel instanceof HzInitStatmt) {
+        this.addStatement(new InitStatementNode(topLevel))
+      } else if (Array.isArray(topLevel)) {
+        for (const varDecl of topLevel) {
+          if (varDecl instanceof HzVarDecl) {
+            this.defineLocal(new LocalVarNode(varDecl.name))
+          }
+        }
+      }
+    }
   }
 }
 
 export abstract class RefNode extends ASTNode {
   name: string
   constant: boolean = false
-  constructor(name: string) {
+  constructor() {
     super()
-    this.name = name
   }
 }
 
 export class LocalVarNode extends RefNode {
+  constructor(name: string) {
+    super()
+    this.name = name
+  }
   get isLeaf(): boolean {
     return true
   }
@@ -76,7 +138,8 @@ export class ParamNode extends LocalVarNode {
 export class FieldNode extends RefNode {
   declare parent: ObjNode
   constructor(name: string) {
-    super(name)
+    super()
+    this.name = name
   }
   get isLeaf(): boolean {
     return true
@@ -88,9 +151,11 @@ export class ObjNode extends RefNode {
   fields: Map<string, FieldNode> = new Map()
   objMethods: Map<string, ObjMethodNode> = new Map()
   classMethods: Map<string, ClassMethodNode> = new Map()
-  constructor(name: string) {
-    super(name)
+  def: HzObjDecl
+  constructor(def: HzObjDecl) {
+    super()
     this.constant = true
+    this.def = def
   }
   findRef(name: string): RefNode | undefined {
     return this.fields.get(name)
@@ -126,13 +191,35 @@ export class ObjNode extends RefNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield* this.fields.values()
+    yield* this.ctors.values()
+    yield* this.classMethods.values()
+    yield* this.objMethods.values()
+  }
+  build(): void {
+    for (const field of this.def.fields) {
+      this.defineField(new FieldNode(field.name))
+    }
+    for (const ctor of this.def.ctors) {
+      this.defineCtor(new CtorNode(ctor))
+    }
+    for (const method of this.def.objMethods) {
+      this.defineObjMethod(new ObjMethodNode(method))
+    }
+    for (const method of this.def.classMethods) {
+      this.defineClassMethod(new ClassMethodNode(method))
+    }
+  }
 }
 
 export class FuncNode extends RefNode {
   params: Map<string, ParamNode> = new Map()
   body: BlockNode
-  constructor(name: string) {
-    super(name)
+  def: HzFuncDecl
+  constructor(def: HzFuncDecl) {
+    super()
+    this.def = def
     this.constant = true
   }
   defParam(param: ParamNode): void {
@@ -153,30 +240,30 @@ export class FuncNode extends RefNode {
   get isLeaf(): boolean {
     return false
   }
-}
-
-export class ClassMethodNode extends FuncNode {
-  declare parent: ObjNode
-  constructor(name: string) {
-    super(name)
-    this.constant = true
+  *children(): Iterable<ASTNode> {
+    yield* this.params.values()
+    yield this.body
+  }
+  build(): void {
+    if (this.def instanceof HzNaryFuncDecl) {
+      for (const { selector, param } of this.def.selectors) {
+        this.defParam(new ParamNode(param ?? selector))
+      }
+    }
+    this.defBody(new BlockNode(this.def.body))
   }
 }
 
 export class CtorNode extends FuncNode {
   declare parent: ObjNode
-  constructor(name: string) {
-    super(name)
-    this.constant = true
-  }
+}
+
+export class ClassMethodNode extends FuncNode {
+  declare parent: ObjNode
 }
 
 export class ObjMethodNode extends FuncNode {
   declare parent: ObjNode
-  constructor(name: string) {
-    super(name)
-    this.constant = true
-  }
 }
 
 export class SelfRefNode extends LocalVarNode {
@@ -193,9 +280,10 @@ export abstract class ExprNode extends ASTNode {
 export class RefExprNode extends ExprNode {
   /** unowned */
   ref: RefNode
-  constructor(ref: RefNode) {
+  protected def: HzRefExpr
+  constructor(def: HzRefExpr) {
     super()
-    this.ref = ref
+    this.def = def
   }
   get isSingle(): boolean {
     return true
@@ -209,9 +297,10 @@ export class BinaryExprNode extends ExprNode {
   left: ExprNode
   op: Operator
   right: ExprNode
-  constructor(op: Operator) {
+  def: HzBinaryExpr
+  constructor(def: HzBinaryExpr) {
     super()
-    this.op = op
+    this.def = def
   }
   get isSingle(): boolean {
     return false
@@ -231,13 +320,27 @@ export class BinaryExprNode extends ExprNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.left
+    yield this.right
+  }
+  build(): void {
+    this.defLeft(makeExprNode(this.def.left))
+    this.op = this.def.op
+    this.defRight(makeExprNode(this.def.right))
+  }
 }
 
 export class FuncCallExprNode extends ExprNode {
   caller?: ExprNode
   /** unowned */
-  func: FuncNode
+  func: FuncNode | string
   args: ExprNode[] = []
+  def: HzFuncCallExpr
+  constructor(def: HzFuncCallExpr) {
+    super()
+    this.def = def
+  }
   get isSingle(): boolean {
     return true
   }
@@ -247,9 +350,10 @@ export class FuncCallExprNode extends ExprNode {
     this.caller = caller
     caller.parent = this
   }
-  setFunc(func: FuncNode) {
+  setFunc(func: FuncNode | string) {
     if (this.func)
-      throw new ASTNodeDefinedError("func already defined in func call", this, func)
+      throw new ASTNodeDefinedError("func already defined in func call", this,
+        func instanceof FuncNode ? func : undefined)
     this.func = func
   }
   addArg(arg: ExprNode) {
@@ -259,38 +363,78 @@ export class FuncCallExprNode extends ExprNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    if (this.caller) yield this.caller
+    yield* this.args.values()
+  }
+  build(): void {
+    if (this.def instanceof HzNullaryFuncCallExpr) {
+      const signature = getFuncSignature(this.def.selector)
+      const func = this.parent?.findRef(signature)
+      this.setFunc(func instanceof FuncNode ? func : signature)
+      if (this.def.caller) {
+        this.setCaller(makeExprNode(this.def.caller))
+      }
+    } else if (this.def instanceof HzNaryFuncCallExpr) {
+      const signature = getFuncSignature(this.def.selectors)
+      const func = this.parent?.findRef(signature)
+      this.setFunc(func instanceof FuncNode ? func : signature)
+      if (this.def.caller) {
+        this.setCaller(makeExprNode(this.def.caller))
+      }
+      for (const arg of this.def.selectors) {
+        this.addArg(makeExprNode(arg.arg))
+      }
+    }
+  }
 }
 
-export class DynamicFuncCallExprNode extends ExprNode {
-  funcName: string
-  caller?: ExprNode
-  args: ExprNode[] = []
-  get isSingle(): boolean {
-    return true
-  }
-  setCaller(caller: ExprNode) {
-    if (this.caller)
-      throw new ASTNodeDefinedError("call already defined in func call", this, caller)
-    this.caller = caller
-    caller.parent = this
-  }
-  addArg(arg: ExprNode) {
-    this.args.push(arg)
-    arg.parent = this
-  }
-  get isLeaf(): boolean {
-    return false
-  }
-}
-
-export class LiteralExprNode<T = unknown> extends ExprNode {
+export class LiteralExprNode extends ExprNode {
+  protected def: HzLiteralExpr
   raw: string
-  value: T
+  value: unknown
+  type: LiteralType
+  constructor(def: HzLiteralExpr) {
+    super()
+    this.def = def
+  }
   get isSingle(): boolean {
     return true
   }
   get isLeaf(): boolean {
     return true
+  }
+  build(): void {
+    const literal = this.def.value
+    this.type = this.def.value.type
+    this.raw = literal.raw
+    if (literal.type === LiteralType.bool) {
+      this.value = literal.raw === SoftKeyword.true
+    } else if (literal.type === LiteralType.string) {
+      this.value = literal.raw
+    } else if (literal.type === LiteralType.number) {
+      this.value = Number(literal.raw)
+    } else if (literal.type === LiteralType.undefined) {
+      this.value = undefined
+    } else if (literal.type === LiteralType.null) {
+      this.value = null
+    }
+  }
+}
+
+function makeExprNode(def: HzExpr): ExprNode {
+  if (def instanceof HzLiteralExpr) {
+    return new LiteralExprNode(def)
+  } else if (def instanceof HzRefExpr) {
+    return new RefExprNode(def)
+  } else if (def instanceof HzNullaryFuncCallExpr) {
+    return new FuncCallExprNode(def)
+  } else if (def instanceof HzNaryFuncCallExpr) {
+    return new FuncCallExprNode(def)
+  } else if (def instanceof HzBinaryExpr) {
+    return new BinaryExprNode(def)
+  } else {
+    throw new SemanticAnalysisError("Unrecognized expr definition", def)
   }
 }
 
@@ -299,11 +443,21 @@ export abstract class StatementNode extends ASTNode {
 }
 
 export class BreakStatementNode extends StatementNode {
+  protected def: HzBreakStatmt
+  constructor(def: HzBreakStatmt) {
+    super()
+    this.def = def
+  }
   get isLeaf(): boolean {
     return true
   }
 }
 export class ContinueStatementNode extends StatementNode {
+  protected def: HzContinueStatmt
+  constructor(def: HzContinueStatmt) {
+    super()
+    this.def = def
+  }
   get isLeaf(): boolean {
     return true
   }
@@ -311,6 +465,11 @@ export class ContinueStatementNode extends StatementNode {
 
 export class ExprStatementNode extends StatementNode {
   expr: ExprNode
+  protected def: HzExprStatmt
+  constructor(def: HzExprStatmt) {
+    super()
+    this.def = def
+  }
   defineExpr(expr: ExprNode): void {
     if (this.expr)
       throw new ASTNodeDefinedError("Expr already defined in statement", this, expr)
@@ -320,25 +479,36 @@ export class ExprStatementNode extends StatementNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.expr
+  }
+  build(): void {
+    this.defineExpr(makeExprNode(this.def.expr))
+  }
 }
 
 export class IfStatementNode extends StatementNode {
   condition: ExprNode
   consequent: BlockNode
   alternate?: BlockNode
-  defineCondition(condition: ExprNode): void {
+  protected def: HzIfStatmt
+  constructor(def: HzIfStatmt) {
+    super()
+    this.def = def
+  }
+  defCondition(condition: ExprNode): void {
     if (this.condition)
       throw new ASTNodeDefinedError("Condition already defined in if", this, condition)
     this.condition = condition
     condition.parent = this
   }
-  defineConsequent(consequent: BlockNode): void {
+  defConsequent(consequent: BlockNode): void {
     if (this.consequent)
       throw new ASTNodeDefinedError("Consequent already defined in if", this, consequent)
     this.consequent = consequent
     consequent.parent = this
   }
-  defineAlternate(alternate: BlockNode): void {
+  defAlternate(alternate: BlockNode): void {
     if (this.alternate)
       throw new ASTNodeDefinedError("Alternate already defined in if", this, alternate)
     this.alternate = alternate
@@ -347,18 +517,37 @@ export class IfStatementNode extends StatementNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.condition
+    yield this.consequent
+    if (this.alternate) yield this.alternate
+  }
+  build(): void {
+    this.defCondition(makeExprNode(this.def.condition))
+    // consequent
+    this.defConsequent(new BlockNode(this.def.consequent))
+    // alternate
+    if (this.def.alternate) {
+      this.defAlternate(new BlockNode(this.def.alternate))
+    }
+  }
 }
 
 export class WhileStatementNode extends StatementNode {
   condition: ExprNode
   body: BlockNode
-  defineCondition(condition: ExprNode): void {
+  protected def: HzWhileStatmt
+  constructor(def: HzWhileStatmt) {
+    super()
+    this.def = def
+  }
+  defCondition(condition: ExprNode): void {
     if (this.condition)
       throw new ASTNodeDefinedError("Condition already defined in while", this, condition)
     this.condition = condition
     condition.parent = this
   }
-  defineBody(body: BlockNode): void {
+  defBody(body: BlockNode): void {
     if (this.body)
       throw new ASTNodeDefinedError("Body already defined in while", this, body)
     this.body = body
@@ -367,10 +556,23 @@ export class WhileStatementNode extends StatementNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.condition
+    yield this.body
+  }
+  build(): void {
+    this.defCondition(makeExprNode(this.def.condition))
+    this.defBody(new BlockNode(this.def.body))
+  }
 }
 
 export class ReturnStatementNode extends StatementNode {
   value: ExprNode
+  def: HzReturnStatmt
+  constructor(def: HzReturnStatmt) {
+    super()
+    this.def = def
+  }
   defValue(value: ExprNode): void {
     if (this.value)
       throw new ASTNodeDefinedError("Value already defined in return", this, value)
@@ -380,12 +582,23 @@ export class ReturnStatementNode extends StatementNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.value
+  }
+  build(): void {
+    this.defValue(makeExprNode(this.def.value))
+  }
 }
 
 export class InitStatementNode extends StatementNode {
   /** unowned */
   lvalue: LocalVarNode
   rvalue: ExprNode
+  protected def: HzInitStatmt
+  constructor(def: HzInitStatmt) {
+    super()
+    this.def = def
+  }
   defRvalue(rvalue: ExprNode): void {
     if (this.rvalue)
       throw new ASTNodeDefinedError("Rvalue already defined in init", this, rvalue)
@@ -395,15 +608,54 @@ export class InitStatementNode extends StatementNode {
   get isLeaf(): boolean {
     return false
   }
+  *children(): Iterable<ASTNode> {
+    yield this.rvalue
+  }
+  build(): void {
+    const lvalueName = this.def.name
+    const lvalue = this.parent?.findRef(lvalueName)
+    if (lvalue === undefined)
+      throw new SemanticAnalysisError(`${lvalue} not declared`, this.def)
+    this.lvalue = lvalue
+    this.defRvalue(makeExprNode(this.def.value))
+  }
+}
+
+function makeStatementNode(def: HzStatmt): StatementNode {
+  if (def instanceof HzIfStatmt) {
+    return new IfStatementNode(def)
+  } else if (def instanceof HzWhileStatmt) {
+    return new WhileStatementNode(def)
+  } else if (def instanceof HzInitStatmt) {
+    return new InitStatementNode(def)
+  } else if (def instanceof HzExprStatmt) {
+    return new ExprStatementNode(def)
+  } else if (def instanceof HzReturnStatmt) {
+    return new ReturnStatementNode(def)
+  } else if (def instanceof HzBreakStatmt) {
+    return new BreakStatementNode(def)
+  } else if (def instanceof HzContinueStatmt) {
+    return new BreakStatementNode(def)
+  } else {
+    throw new SemanticAnalysisError("Unrecognized statement definition", def)
+  }
 }
 
 export class ASTNodeDefinedError extends Error {
   parent: ASTNode
-  node: ASTNode
-  constructor(message: string, parent: ASTNode, node: ASTNode) {
+  node?: ASTNode
+  constructor(message: string, parent: ASTNode, node?: ASTNode) {
     super(message)
     this.name = this.constructor.name
     this.parent = parent
     this.node = node
+  }
+}
+
+export class SemanticAnalysisError extends Error {
+  constructor(message: string, cause: unknown) {
+    super(message)
+    this.name = this.constructor.name
+    this.cause = cause
   }
 }
