+++ src/index.ts
@@ -1,6 +1,6 @@
 import { Writable } from "stream"
-import { SemanticAnalyzeError, semanticAnalyze } from "./ast/analysis.js"
-import { ASTNodeDefinedError } from "./ast/node.js"
+import { semanticAnalyze } from "./ast/analysis.js"
+import { ASTNodeDefinedError, SemanticAnalysisError } from "./ast/node.js"
 import { lex } from "./lex/lexer.js"
 import { ParseError, parse } from "./parse/parser.js"
 import { transpile2Js } from "./visitor/2js.js"
@@ -64,7 +64,7 @@ try {
     console.log(nearby(source, token.pos))
   } else if (e instanceof ASTNodeDefinedError) {
     //
-  } else if (e instanceof SemanticAnalyzeError) {
+  } else if (e instanceof SemanticAnalysisError) {
     //
   }
   console.error(e, full)
